backend/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
Django==4.2
django-graphql-geojson==0.1.4
geographiclib==2.0
geopy==2.3.0
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
ordered-set==4.1.0
promise==2.3
psycopg==3.1.8
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.5.0
